This was my first day of using the HP TestDrive set of computers. Here is what I accomplished today:
1.  Got back in working order the HP/UX port using an HP TestDrive system
    running HP/UX 11 on an HP 9000/800 series computer.

2.  Fixed a bunch of printf() and scanf() family format string errors using
    the Intel C compiler version 9.0 on Red Hat Enterprise Linux AS 4.0.



git-svn-id: http://svn.csrri.iit.edu/mx/trunk@239 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_gm10_timer.c
@@ -365,9 +365,9 @@ mxd_gm10_timer_is_busy( MX_TIMER *timer )
 
 #if MXD_GM10_TIMER_DEBUG
 
-	MX_DEBUG(-2,("%s: current time = (%lu,%ld), busy = %d",
+	MX_DEBUG(-2,("%s: current time = (%lu,%lu), busy = %d",
 		fname, current_time_in_clock_ticks.high_order,
-		(unsigned long) current_time_in_clock_ticks.low_order,
+		current_time_in_clock_ticks.low_order,
 		timer->busy ));
 #endif
