+++ libMx/d_smartmotor_aio.c
@@ -291,7 +291,7 @@ mxd_smartmotor_ain_finish_record_initialization( MX_RECORD *record )
 
 	for ( i = 0; i < length; i++ ) {
 		if ( islower( (int)(port_name[i]) ) ) {
-			port_name[i] = toupper( port_name[i] );
+			port_name[i] = toupper( (int)(port_name[i]) );
 		}
 	}
 
@@ -429,7 +429,7 @@ mxd_smartmotor_aout_finish_record_initialization( MX_RECORD *record )
 
 	for ( i = 0; i < length; i++ ) {
 		if ( islower( (int)(port_name[i]) ) ) {
-			port_name[i] = toupper( port_name[i] );
+			port_name[i] = toupper( (int)(port_name[i]) );
 		}
 	}
